+++ lib/shopify_api/webhooks/registrations/http.rb
@@ -36,6 +36,8 @@ module ShopifyAPI
                 edges {
                   node {
                     id
+                    includeFields
+                    metafieldNamespaces
                     endpoint {
                       __typename
                       ... on WebhookHttpEndpoint {
@@ -49,10 +51,19 @@ module ShopifyAPI
           QUERY
         end
 
-        sig { override.params(body: T::Hash[String, T.untyped]).returns(T::Hash[Symbol, String]) }
+        sig do
+          override.params(body: T::Hash[String, T.untyped]).returns({
+            webhook_id: T.nilable(String),
+            current_address: T.nilable(String),
+            fields: T::Array[String],
+            metafield_namespaces: T::Array[String],
+          })
+        end
         def parse_check_result(body)
           edges = body.dig("data", "webhookSubscriptions", "edges") || {}
           webhook_id = nil
+          fields = []
+          metafield_namespaces = []
           current_address = nil
           unless edges.empty?
             node = edges[0]["node"]
@@ -63,8 +74,11 @@ module ShopifyAPI
               else
                 node["callbackUrl"].to_s
               end
+            fields = node["includeFields"] || []
+            metafield_namespaces = node["metafieldNamespaces"] || []
           end
-          { webhook_id: webhook_id, current_address: current_address }
+          { webhook_id: webhook_id, current_address: current_address, fields: fields,
+            metafield_namespaces: metafield_namespaces, }
         end
       end
     end